requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
gunicorn==23.0.0
joblib==1.4.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
databricks-sql-connector==4.0.3
python-dotenv==1.1.0
mlflow==2.22.0